Add favorite feature and product detail screen
Implemented favorite API integration with repository and UI, including toggling favorites in HomeScreen and HomeViewModel. Added ProductScreen and ProductViewModel for detailed product view, including related products and seller info. Updated ProductDTO price type to Double and improved utility functions for date formatting. Enhanced ProductCard to handle missing seller avatars. (van con loi dieu huong)

+++ app/src/main/java/com/chosv/chosv_android/data/model/ProductDTO.kt
@@ -24,7 +24,7 @@ data class ProductDetail(
     val sellerAddress: String,
     val sellerPhone: String,
     val productDescription: String,
-    val price: Long,
+    val price: Double, // Sửa từ Long thành Double
     val status: String,
     val createdDate: String,
     val sellerFullName: String,
